+++ lang/en/block_people.php
@@ -31,5 +31,19 @@ $string['people:myaddinstance'] = 'Add a new people block to Dashboard';
 $string['noparticipantslist'] = 'Viewing the participants list is prohibited in this course';
 
 // Settings.
+$string['setting_participantspageheading'] = 'Participants page';
 $string['setting_linkparticipantspage'] = 'Show link to the participants page';
 $string['setting_linkparticipantspage_desc'] = 'By enabling this setting, a link to the participants page of the course will be shown within the block.';
+$string['setting_hideblockheading'] = 'Hiding the block';
+$string['setting_hideblock'] = 'Hiding the block';
+$string['setting_hideblock_desc'] = 'By enabling this setting, the block can be hidden by users.<br/>
+Important notice:<br/>
+Disabling this setting will entirely remove the showing / hiding the block menu item. This means, that users cannot hide this block anymore, but on the other hand, blocks that are already hidden cannot be shown anymore, too. If you want to enable this feature, consider using the following function to reset the visibility for all "block_people" instances.';
+$string['setting_resetvisibility'] = 'Reset visibility';
+$string['setting_resetvisibility_desc'] = 'By enabling this checkbox, the visibility of all existing "block_people" instances will be set to visible (again).<br/>
+Please note: <br/>
+After saving this option, the database operations for resetting the visibility will be triggered and this checkbox will be unticked again. The next enabling and saving of this feature will trigger the database operations for resetting the visibility again. ';
+
+// Notifications.
+$string['resetvisibilitysuccess'] = 'Success! All "block_people" instances are visible (again). <br/> The setting "Reset visibility" has been reset.';
+$string['resetvisibilityerror'] = 'Oops... Something went wrong updating the database tables... <br/> The setting "Reset visibility" has been reset.';